requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
notion-client>=1.0.0
aiohttp>=3.8.0
python-dateutil>=2.8.2
typing-extensions>=4.0.0
asyncio>=3.4.3 